% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HideyHole.R
\name{HideyHole}
\alias{HideyHole}
\title{HideyHole - Look for small holes in a digital elevation model raster}
\usage{
HideyHole(
  r,
  neighbourhood = 21,
  hole.depth = 0.1,
  min.pixels = 1,
  max.pixels = NA
)
}
\arguments{
\item{r}{terra::rast object, a dtm grid}

\item{neighbourhood}{diamter in grid pixels for searching for holes}

\item{hole.depth}{depth in elevation units (typically m) to define local holes / depressions (default 21)}

\item{min.pixels}{minimum size of hole (in number of pixels), (default 100)}

\item{max.pixels}{maximum size of hole (in number of pixels), default NA = no maximum}
}
\value{
a list of the following elements
HideyHoleCount: Number of distinct holes found
HideyHoleCoverage: Proportion (0-1) of grid covered by hidey holes (2D area of holes / 2d area of grid)
HideyHoleVector: SpatVector polygon object of each hole with attributes of area, pixels & deepest point (difference from local average)
HideyHoleRaster: SpatRaster object a binary raster grid with 1=hidey hole, NA otherwise
neighbourhood: Input parameter value
hole.depth: Input parameter value
min.pixels: Input parameter value
max.pixels: Input parameter value
}
\description{
HideyHole - Look for small holes in a digital elevation model raster
}
\details{
HideyHole() find local depressions or holes in a digital elevation model that could be hiding spaces for organisms
such as fish or crustaceans. Achieved by comparing pixel heights with local averages, withs holes defined as clusters of pixels
deeper than the local average by a given measure.
}
\examples{
TestGrid<-terra::rast("https://github.com/cyesson/HideyHole/raw/main/data/TestGrid.tif")
HideyHole(TestGrid, neighbourhood=5)
}
